requirements.txt

djangorestframework
djangorestframework-simplejwt
PyJWT
qrcode
Pillow
python-dotenv
